+++ tests/fixtures/noqa/noqa.py
@@ -1,5 +1,4 @@
-#!/usr/bin/env pytho # noqa: WPS453 WPS323
-# -*- coding: utf-8 -*-
+# -*- coding: utf-8 -*- # noqa: WPS323
 """
 This file contains all possible violations.