+++ tests/fixtures/test_executable_mismatch_resources/exe001_neg.py
@@ -1,3 +1,0 @@
-if __name__ == '__main__':
-    print('I should be executable.')
-# flake8: noqa