+++ tests/fixtures/test_executable_mismatch_resources/exe001_pos.py
@@ -1,5 +1,0 @@
-#!/usr/bin/python # flake8: noqa
-
-if __name__ == '__main__':
-    print('I should be executable.')
-# flake8: noqa