+++ tests/fixtures/test_executable_mismatch_resources/exe002_neg.py
@@ -1,3 +1,0 @@
-def a_lib_function():
-    print("I shouldn't be executable.")
-# flake8: noqa