+++ tests/fixtures/test_executable_mismatch_resources/exe003_neg.py
@@ -1,5 +1,0 @@
-#!/usr/bin/python3
-
-if __name__ == '__main__':
-    print('Wrong shebang.')
-# flake8: noqa