+++ tests/fixtures/test_executable_mismatch_resources/exe003_pos.py
@@ -1,5 +1,0 @@
-#!/bin/bash # flake8: noqa
-
-if __name__ == '__main__':
-    print('Wrong shebang.')
-# flake8: noqa