+++ tests/fixtures/test_executable_mismatch_resources/exe004_neg.py
@@ -1,5 +1,0 @@
-#!/usr/bin/python3
-
-if __name__ == '__main__':
-    print('I do not have whitespace before shebang.')
-# flake8: noqa