+++ tests/fixtures/test_executable_mismatch_resources/exe005_pos.py
@@ -1,7 +1,0 @@
-
-#
-#!/usr/bin/python3
-
-if __name__ == '__main__':
-    print('I have blank and comment lines before shebang.')
-# flake8: noqa