+++ tests/plugins/tokenize_parser.py
@@ -17,11 +17,10 @@ def parse_tokens():
 
 
 @pytest.fixture(scope='session')
-def parse_file_tokens():
+def parse_file_tokens(parse_tokens):
     """Parses tokens from a file."""
     def factory(filename: str):
         with open(filename, 'r', encoding='utf-8') as test_file:
             file_content = test_file.read()
-            lines = io.StringIO(dedent(file_content))
-            return list(tokenize.generate_tokens(lambda: next(lines)))
+            return parse_tokens(file_content)
     return factory